Fix up rename leftovers: test include paths, README, CI action
Step 1 renamed the directories but the test sources and external
references still pointed at the old paths.

Test sources:
 - All 9 test files (MockMainWindowView.h + 8 test_*.cpp) had
   `#include "../GhosttyWin32App/..."` lines from the original
   GhosttyTests layout. Switched to `../App/...` so the test EXE
   picks up the renamed App headers.

README:
 - Path references for `GhosttyWin32App.exe`, build dirs, etc.
   The advertised EXE name is `GhosttyWin32.exe` (TargetName
   intentionally stays GhosttyWin32; see step 1 commit), not
   `App.exe`.

CI action (build-signed-msix):
 - nuget restore lines now hit App/App.vcxproj and
   Tests/Tests.vcxproj.
 - manifest / wapproj / AppPackages paths under Package/.
