+++ .github/actions/build-signed-msix/action.yml
@@ -68,7 +68,7 @@ runs:
         cd ghostty
         zig build -Doptimize=ReleaseSafe -Drenderer=directx
         # upstream renamed the install artifact: zig-out/lib/ghostty.dll →
-        # ghostty-internal.dll (ghostty commit 4fd16ef9b). GhosttyWin32App
+        # ghostty-internal.dll (ghostty commit 4fd16ef9b). App
         # still links against ghostty.dll, so rename on copy.
         Copy-Item zig-out/lib/ghostty-internal.dll ../ghostty/ghostty.dll
         Copy-Item include/ghostty.h ../ghostty/
@@ -97,7 +97,7 @@ runs:
       uses: NuGet/setup-nuget@v2
 
     # The wapproj uses PackageReference (handled by msbuild /t:Restore),
-    # but GhosttyWin32App.vcxproj / GhosttyTests.vcxproj still use the
+    # but App.vcxproj / Tests.vcxproj still use the
     # legacy packages.config format which msbuild's Restore target ignores.
     # We need both: nuget restore expands ..\packages\ for the vcxprojs,
     # then msbuild /t:Restore resolves the wapproj's PackageReferences.
@@ -107,8 +107,8 @@ runs:
       run: |
         $ErrorActionPreference = 'Stop'
         $PSNativeCommandUseErrorActionPreference = $true
-        nuget restore "GhosttyWin32App/GhosttyWin32App.vcxproj" -PackagesDirectory packages -SolutionDirectory .
-        nuget restore "GhosttyTests/GhosttyTests.vcxproj" -PackagesDirectory packages -SolutionDirectory .
+        nuget restore "App/App.vcxproj" -PackagesDirectory packages -SolutionDirectory .
+        nuget restore "Tests/Tests.vcxproj" -PackagesDirectory packages -SolutionDirectory .
         msbuild GhosttyWin32.slnx /t:Restore /p:Configuration=Release /p:Platform=x64
 
     - name: Patch manifest version
@@ -125,7 +125,7 @@ runs:
         # Use -creplace (case-sensitive) on uppercase Version=, otherwise
         # -replace also clobbers `version="1.0"` in the <?xml ... ?>
         # declaration since PowerShell -replace is case-insensitive.
-        $manifestPath = "GhosttyWin32 (Package)/Package.appxmanifest"
+        $manifestPath = "Package/Package.appxmanifest"
         $content = Get-Content $manifestPath -Raw
         # (?<=\s) lookbehind requires whitespace before Version=, so
         # MinVersion= / MaxVersionTested= in <TargetDeviceFamily> aren't
@@ -176,7 +176,7 @@ runs:
         $ErrorActionPreference = 'Stop'
         $PSNativeCommandUseErrorActionPreference = $true
         $msbuildArgs = @(
-          "GhosttyWin32 (Package)/GhosttyWin32 (Package).wapproj",
+          "Package/Package.wapproj",
           "/p:Configuration=Release",
           "/p:Platform=x64",
           "/p:PlatformToolset=v143",
@@ -211,11 +211,11 @@ runs:
       run: |
         $ErrorActionPreference = 'Stop'
         $PSNativeCommandUseErrorActionPreference = $true
-        $msix = Get-ChildItem -Recurse -Filter "*.msix" -Path "GhosttyWin32 (Package)/AppPackages" |
+        $msix = Get-ChildItem -Recurse -Filter "*.msix" -Path "Package/AppPackages" |
                 Where-Object { $_.Name -notlike "*Microsoft.*" } |
                 Select-Object -First 1
         if (-not $msix) {
-          Write-Error "Built MSIX not found under GhosttyWin32 (Package)/AppPackages/"
+          Write-Error "Built MSIX not found under Package/AppPackages/"
           exit 1
         }
         Write-Host "MSIX: $($msix.FullName)"
